Make a crude downloading tool with Cocoa
Create a Foundation tool that downloads a single URL (via command-line
argument) to the working directory with an automatically-defined
file-name. Downloading uses the NSURLSession API with bare-bones
settings and a basic run loop. Error handling is also basic.

Add an Info.plist file that is merged into the flat-file product.