Change error messages
In the download task completion handler, limit printed error messages
to just the localized description, removing the localized failure
reason. I found a bug (<rdar://problem/20367692>) with one failure
reason, and it makes the text too long. (It may sometimes be included
in a synthesized description, leading to doubled text.)

Surround an expression for a method argument with parentheses.

Bump up build number.