Move globals to main; customize error codes
Since the variables for the return code and the exit flag are used only
in the main function or a block within the main function, move both
variables to inside the main function.

Instead of a single command-line return value for errors, set up a list
of (documented) potential values for differing causes.