Remove buggy code
For the helper function that moves a file to a given volume’s
item-replacement directory, remove the code that skips the move if the
source file is already on the destination volume. (Now the move is
unconditional.) When the move is skipped, sometimes NSURLSession’s
temporary-file shredder is triggered early.

Add documentation for the helper functions.