Add command-line options for printing version and help strings
Add category on the GBCli settings class for program-specific options.
Add new class with methods to return GBCli settings and options objects
with program-specific data. Add key strings for the options and
settings domains.

Replace command-line processing for no arguments or a URL argument with
GBCli objects. Add option to print the tool version and exit. Add
option to print the help text and exit.

Add error return code when a print-and-exit option and a URL are both
submitted.

Bump up build number.