Add option to not print the failed-download marker
Add option to not print a line containing “#” when a URL download
fails, opposed to the new file’s path when it succeeds. This results to
a clean list of URLs going to standard output at the cost of not
matching input URLs to output lines. The option phrase is negative
since having an option defaulting to ON and needing the “—no-*” variant
to override is currently broken in GBCli
(<https://github.com/tomaz/GBCli/issues/16>). Change the branch logic
of printing the result URLs/error-marker and error messages.

Change the help text to indicate at including at least one URL is
mutually exclusive with printing the help and/or version text. (Other
options may mix with submitting URLs.)

Bump up the build number.