Add option to submit URLs via an input file
Add option to read (additional) URLs from a given file (if the option
includes a file path) or standard input (if the option is bare). Note
in the help text that there may be zero URLs if the input file option
is used. However, ensure that printing the help and/or version text
while listing URLs, from either the command line or an input file, is
still an error. Add return codes for not being able to open or read an
input file.

Add code to remove a session object early if there are no owned tasks
(since the session’s delete command is usually sent when the last task
finishes).

Bump up build number.