Add options to redirect downloaded files
Add option to specify a path where downloaded files will go, and other
option specifying if the path is to a file or directory. The second
option is valid only if the first is given. If not given, the first
option defaults to the current working directory, and the second to a
file if one URL is given and a directory if multiple URLs are given.
Add return code for when a destination directory could not be created
or made the working directory. (Note while using the same name for
multiple files, including a pre-existing file object, will invoke
contingency renaming, specifying a directory will fail if there’s an
existing non-directory with that name.)

Add names for each group of options in the help text.

Bump up build number.